requirements.txt
Flask==0.12.2
pycrypto==2.6.1
flask_jwt_extended==3.24.0
flask_mongoengine==0.9.5
Unidecode==1.1.1
